Fix type issues, create SelectOptionBox file
